Implement 3_x86_32, 3_x86_128, 3_x64_128

+++ ext/digest/murmurhash/3_x86_32.c
@@ -74,7 +74,8 @@ VALUE
 murmur3_x86_32_s_digest(int argc, VALUE *argv, VALUE klass)
 {
   uint8_t digest[4];
-  uint64_t h;
+  uint32_t h;
+
   h = _murmur_s_digest32(argc, argv, klass, murmur_hash_process3_x86_32);
   assign_by_endian_32(digest, h);
   return rb_str_new((const char*) digest, 4);
